load commands from src rather than config
