+++ src/__tests__/index.test.js
@@ -11,88 +11,32 @@ function uncacheIndex() {
 }
 
 describe(testContext(__filename), function () {
-  describe('imports and exports', function () {
-    before(function () {
-      uncacheIndex()
-      this.commands = require('../index')
-    })
-
-    it('usage(), parse(), and commandDescriptor exist for every command', function () {
-      Object.keys(this.commands).forEach(cmdName => {
-        const command = this.commands[cmdName]
-        expect(command.commandDescriptor).to.have.property('name')
-        expect(typeof command.usage).to.equal('function')
-        expect(typeof command.parse).to.equal('function')
-      })
-    })
-  })
-
   describe('ignore inactive', function () {
     before(function () {
-      this.cmdConfigFilePath = cmdName => path.resolve(__dirname, '..', '..', 'config', 'commands', `${cmdName}.yaml`)
       this.cmdSourceFilePath = cmdName => path.resolve(__dirname, '..', 'commands', `${cmdName}.js`)
       this.testCommandsToWrite = {
-        testcmd1: {
-          config: `%YAML 1.2
----
-command:
-  name: testcmd1
-  description: generated by tests (2)
-  usage: testcmd1 [options] <command>
-  _inactive: true
-`,
-
-          source: `
+        testcmd1: `
 export function invoke(argv) {
   return false
 }
-`
-        },
-
-        testcmd2: {
-          config: `%YAML 1.2
----
-command:
-  name: testcmd2
-  description: generated by tests (1)
-  usage: testcmd2 [options] <command>
-  commands:
-    -
-      name: present
-      description: this subcommand should be present
-      usage: present [options]
-    -
-      name: ignored
-      description: this subcommand should be ignored
-      usage: launch [options]
-      _inactive: true
-  `,
-
-          source: `
-export function invoke(argv) {
-  return false
+export const commandDescriptor = {
+  name: 'testcmd1',
+  description: 'test command 1',
+  _inactive: true,
 }
-`
-        },
-
-        testcmd3: {
-          config: `%YAML 1.2
----
-command:
-  name: testcmd3
-  description: generated by tests (3)
-  usage: testcmd3 [options] <command>
-  _inactive: true
 `,
 
-          source: 'export const invoke = "NOT A FUNCTION"'
-        },
-
+        testcmd2: `
+export const invoke = 'NOT A FUNCTION'
+export const commandDescriptor = {
+  name: 'testcmd2',
+  description: 'test command 2',
+}
+`,
       }
 
       Object.keys(this.testCommandsToWrite).forEach(cmdName => {
-        fs.writeFileSync(this.cmdConfigFilePath(cmdName), this.testCommandsToWrite[cmdName].config)
-        fs.writeFileSync(this.cmdSourceFilePath(cmdName), this.testCommandsToWrite[cmdName].source)
+        fs.writeFileSync(this.cmdSourceFilePath(cmdName), this.testCommandsToWrite[cmdName])
       })
 
       uncacheIndex()
@@ -101,32 +45,16 @@ command:
 
     after(function () {
       Object.keys(this.testCommandsToWrite).forEach(cmdName => {
-        fs.unlinkSync(this.cmdConfigFilePath(cmdName))
         fs.unlinkSync(this.cmdSourceFilePath(cmdName))
       })
     })
 
     it('ignores commands with _inactive attribute set to true', function () {
       expect(this.requiredCommands).to.not.have.property('testcmd1')
-      expect(this.requiredCommands).to.have.property('testcmd2')
-    })
-
-    it('ignores subcommands with _inactive attribute set to true', function () {
-      expect(this.requiredCommands).to.have.property('testcmd2')
-      const foundIgnored = this.requiredCommands.testcmd2.commandDescriptor.commands.reduce((found, command) => {
-        return found || command.name === 'ignored'
-      }, false)
-      const foundPresent = this.requiredCommands.testcmd2.commandDescriptor.commands.reduce((found, command) => {
-        return found || command.name === 'present'
-      }, false)
-      expect(foundIgnored).to.not.be.ok
-      expect(foundPresent).to.be.ok
     })
 
     it('ignores commands with no invoke() function', function () {
-      expect(this.requiredCommands).not.to.have.property('testcmd1')
-      expect(this.requiredCommands).to.have.property('testcmd2')
-      expect(this.requiredCommands).not.to.have.property('testcmd3')
+      expect(this.requiredCommands).not.to.have.property('testcmd2')
     })
   })
 })
